move shared code out of checkout into shared folder
